refactor(frontend): flatten sdoc page directories
