+++ frontend/config/webpack.entry.js
@@ -10,8 +10,8 @@ const entryFiles = {
   wiki2: '/wiki2.js',
   fileHistory: '/file-history.js',
   fileHistoryOld: '/file-history-old.js',
-  sdocFileHistory: '/pages/sdoc/sdoc-file-history/index.js',
-  sdocPublishedRevision: '/pages/sdoc/sdoc-published-revision/index.js',
+  sdocFileHistory: '/pages/sdoc-file-history/index.js',
+  sdocPublishedRevision: '/pages/sdoc-published-revision/index.js',
   sdocThumbnail: '/pages/sdoc-thumbnail',
   app: '/app.js',
   sharedDirView: '/shared-dir-view.js',
