+++ frontend/src/pages/sdoc-editor/sdoc-chat-panel.css
