+++ frontend/src/pages/sdoc-editor/sdoc-chat-panel.js
@@ -1,13 +1,13 @@
 import React, { useCallback, useLayoutEffect, useMemo } from 'react';
 import { Tooltip } from '@seafile/sdoc-editor';
 import PropTypes from 'prop-types';
-import DirChat from '../../../components/dir-view-mode/dir-chat';
-import { ASK_PAGE_SLUG_ID } from '../../../components/dir-view-mode/dir-chat/constants';
-import { AttachmentObject } from '../../../components/dir-view-mode/dir-chat/models';
-import Icon from '../../../components/icon';
-import OpIcon from '../../../components/op-icon';
-import ChatToolbar from '../../../components/toolbar/chat-toolbar';
-import { gettext } from '../../../utils/constants';
+import DirChat from '../../components/dir-view-mode/dir-chat';
+import { ASK_PAGE_SLUG_ID } from '../../components/dir-view-mode/dir-chat/constants';
+import { AttachmentObject } from '../../components/dir-view-mode/dir-chat/models';
+import Icon from '../../components/icon';
+import OpIcon from '../../components/op-icon';
+import ChatToolbar from '../../components/toolbar/chat-toolbar';
+import { gettext } from '../../utils/constants';
 import { getDocSessionIds, getLatestDocSessionId, removeDocSession, removeDocSessions, touchDocSession } from './sdoc-chat-session-storage';
 
 import './sdoc-chat-panel.css';
