+++ frontend/src/pages/sdoc-file-history/history-version.js
@@ -3,13 +3,13 @@ import { Modal, ModalBody } from 'reactstrap';
 import classnames from 'classnames';
 import dayjs from 'dayjs';
 import PropTypes from 'prop-types';
-import CustomDropdown from '../../../components/dropdown';
-import Rename from '../../../components/rename';
-import { gettext, filePath } from '../../../utils/constants';
-import URLDecorator from '../../../utils/url-decorator';
-import { isMobile } from '../../../utils/utils';
+import CustomDropdown from '../../components/dropdown';
+import Rename from '../../components/rename';
+import { gettext, filePath } from '../../utils/constants';
+import URLDecorator from '../../utils/url-decorator';
+import { isMobile } from '../../utils/utils';
 
-import '../../../css/history-record-item.css';
+import '../../css/history-record-item.css';
 
 dayjs.locale(window.app.config.lang);
