+++ frontend/src/pages/sdoc-file-history/side-panel.js
@@ -2,13 +2,13 @@ import React, { Component, Fragment } from 'react';
 import classnames from 'classnames';
 import dayjs from 'dayjs';
 import PropTypes from 'prop-types';
-import Loading from '../../../components/loading';
-import Switch from '../../../components/switch';
-import toaster from '../../../components/toast';
-import { gettext, historyRepoID, PER_PAGE } from '../../../utils/constants';
-import editUtilities from '../../../utils/editor-utilities';
-import { seafileAPI } from '../../../utils/seafile-api';
-import { Utils } from '../../../utils/utils';
+import Loading from '../../components/loading';
+import Switch from '../../components/switch';
+import toaster from '../../components/toast';
+import { gettext, historyRepoID, PER_PAGE } from '../../utils/constants';
+import editUtilities from '../../utils/editor-utilities';
+import { seafileAPI } from '../../utils/seafile-api';
+import { Utils } from '../../utils/utils';
 import { getCurrentAndLastVersion, getLastVersion } from './helper';
 import HistoryVersion from './history-version';
