+++ frontend/src/pages/sdoc-published-revision/index.js
@@ -2,10 +2,10 @@ import React, { Suspense } from 'react';
 import { I18nextProvider } from 'react-i18next';
 import { PublishedRevisionViewer } from '@seafile/seafile-sdoc-editor';
 import { createRoot } from 'react-dom/client';
-import i18n from '../../../_i18n/i18n-sdoc-editor';
-import Loading from '../../../components/loading';
-import { mediaUrl } from '../../../utils/constants';
-import { Utils } from '../../../utils/utils';
+import i18n from '../../_i18n/i18n-sdoc-editor';
+import Loading from '../../components/loading';
+import { mediaUrl } from '../../utils/constants';
+import { Utils } from '../../utils/utils';
 
 const { serviceURL, avatarURL, siteRoot, lang } = window.app.config;
 const { username, name } = window.app.userInfo || {};
